feat(internal): implement Bento APIs
